+++ src/components/input/input.ts
@@ -156,6 +156,14 @@ export default class SlInput extends LitElement {
     this.invalid = !this.input.checkValidity();
   }
 
+  updated(changedProps: Map<string, any>) {
+    // Disabled form controls are always valid, so we need to recheck validity when the state changes
+    if (changedProps.get('disabled')) {
+      this.input.disabled = this.disabled;
+      this.invalid = !this.input.checkValidity();
+    }
+  }
+
   disconnectedCallback() {
     super.disconnectedCallback();
     this.shadowRoot!.removeEventListener('slotchange', this.handleSlotChange);